Calculate metric using overlapping amplified/deleted regions between two samples.

Calculate a specific metric using overlapping amplified/deleted regions between two samples.

calculateOneOverlapMetric(sample01, sample02, method, type)

Arguments

sample01
a GRanges which contains a collection of genomic ranges representing copy number events for the first sample. The GRanges must have a metadata column called 'state' with amplified regions identified as 'AMPLIFICATION' and deleted regions identified as 'DELETION'; regions with different identifications will not be used in the calculation of the metric.

sample02
a GRanges which contains a collection of genomic ranges representing copy number events for the second sample.

method
a character string representing the metric to be used ('sorensen' or 'szymkiewicz'.

type
a character string representing the type of copy number events to be used ('AMPLIFICATION' or 'DELETION').

Value

a numeric, the value of the specified metric. If the metric cannot be calculated, NA is returned.

Details

The method calculates a specified metric using overlapping regions between the samples. Only regions corresponding to the type specified by user are used in the calculation of the metric. The strand of the regions is not taken into account while calculating the metric.

The Sorensen metric is calculated by dividing twice the size of the intersection by the sum of the size of the two sets. If the sum of the size of the two sets is zero; the value NA is returned instead.

The Szymkiewicz-Simpson metric is calculated by dividing the size of the intersection by the smaller of the size of the two sets. If one sample has a size of zero, the metric is not calculated; the value NA is returned instead.

References

Sørensen, Thorvald. n.d. “A Method of Establishing Groups of Equal Amplitude in Plant Sociology Based on Similarity of Species and Its Application to Analyses of the Vegetation on Danish Commons.” Biologiske Skrifter, no. 5: 1–34.

Vijaymeena, M. K, and Kavitha K. 2016. “A Survey on Similarity Measures in Text Mining.” Machine Learning and Applications: An International Journal 3 (1): 19–28. doi: https://doi.org/10.5121/mlaij.2016.3103

Author

Astrid Deschênes

Examples

## Load required package to generate the two samples
require(GenomicRanges)

#> Loading required package: GenomicRanges

#> Loading required package: stats4

#> Loading required package: BiocGenerics

#> Loading required package: parallel

#> 
#> Attaching package: ‘BiocGenerics’

#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from ‘package:base’:
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colnames,
#>     dirname, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, Map, mapply, match, mget,
#>     order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rownames, sapply, setdiff, sort, table, tapply,
#>     union, unique, unsplit, which.max, which.min

#> Loading required package: S4Vectors

#> 
#> Attaching package: ‘S4Vectors’

#> The following object is masked from ‘package:base’:
#> 
#>     expand.grid

#> Loading required package: IRanges

#> Loading required package: GenomeInfoDb

## Generate two samples with identical sequence levels
sample01 <- GRanges(seqnames = "chr1", 
    ranges =  IRanges(start = c(100, 201, 400), 
    end = c(200, 350, 500)), strand =  "*",
    state = c("AMPLIFICATION", "AMPLIFICATION", "DELETION"))
sample02 <- GRanges(seqnames = "chr1", 
    ranges =  IRanges(start = c(150, 200, 450), 
    end = c(250, 350, 500)), strand =  "*",
    state = c("AMPLIFICATION", "DELETION", "DELETION"))

## Calculate Sorensen metric for the amplified regions   
CNVMetrics:::calculateOneOverlapMetric(sample01, sample02, 
    method="sorensen", type="AMPLIFICATION")

#> [1] 0.5738636

## Calculate Szymkiewicz-Simpson metric for the amplified regions   
## Amplified regions of sample02 are a subset of the amplified 
## regions in sample01
CNVMetrics:::calculateOneOverlapMetric(sample01, sample02, 
    method="szymkiewicz", type="AMPLIFICATION")

#> [1] 1

## Calculate Sorensen metric for the deleted regions   
CNVMetrics:::calculateOneOverlapMetric(sample01, sample02, 
    method="sorensen", type="DELETION")

#> [1] 0.3366337

    
## Calculate Szymkiewicz-Simpson metric for the deleted regions    
CNVMetrics:::calculateOneOverlapMetric(sample01, sample02,
    method="szymkiewicz", type="DELETION")

#> [1] 0.5049505