Calculate metric using overlapping amplified/deleted regions

Calculate a specific metric using overlapping amplified/deleted regions between to samples. The metric is calculated for the amplified and deleted regions separately. When more than 2 samples are present, the metric is calculated for each sample pair.

calculateOverlapMetric(segmentData, method = c("sorensen", "szymkiewicz"))

Arguments

segmentData
a GRangesList that contains a collection of genomic ranges representing copy number events, including amplified/deleted status, from at least 2 samples. All samples must have a metadata column called 'state' with amplified regions identified as 'AMPLIFICATION' and deleted regions identified as 'DELETION'; regions with different identifications will not be used in the calculation of the metric.

method
a character string representing the metric to be used. This should be (an unambiguous abbreviation of) one of "sorensen" or "szymkiewicz". Default: "sorensen".

Value

an object of class "CNVMetric" which contains the calculated metric. This object is a list with the following components:

AMPLIFICATION a lower-triangular matrix with the results of the selected metric on the amplified regions for each paired samples. The value NA is present when the metric cannot be calculated. The value NA is also present in the top-triangular section, as well as the diagonal, of the matrix.

DELETION a lower-triangular matrix with the results of the selected metric on the deleted regions for each paired samples. The value NA is present when the metric cannot be calculated. The value NA is also present in the top-triangular section, as well as the diagonal, of the matrix.

The object has the following attributes (besides "class" equal to "CNVMetric"):

metric the metric used for the calculation.

names the names of the two matrix containing the metrics for the amplified and deleted regions.

Details

The two methods each estimate the overlap between paired samples. They use different metrics, all in the range [0, 1] with 0 indicating no overlap.

The available metrics are (written for two GRanges):

sorensen:

This metric is calculated by dividing twice the size of the intersection by the sum of the size of the two sets. With this metric, an overlap metric value of 1 is only obtained when the two samples are identical.

szymkiewicz:

This metric is calculated by dividing the size of the intersection by the size of the smallest set. With this metric, if one set is a subset of the other set, the overlap metric value is 1.

References

Sørensen, Thorvald. n.d. “A Method of Establishing Groups of Equal Amplitude in Plant Sociology Based on Similarity of Species and Its Application to Analyses of the Vegetation on Danish Commons.” Biologiske Skrifter, no. 5: 1–34.

Vijaymeena, M. K, and Kavitha K. 2016. “A Survey on Similarity Measures in Text Mining.” Machine Learning and Applications: An International Journal 3 (1): 19–28. doi: https://doi.org/10.5121/mlaij.2016.3103

Author

Astrid Deschênes, Pascal Belleau

Examples

## Load required package to generate the samples
require(GenomicRanges)

## Create a GRangesList object with 3 samples
## The stand of the regions doesn't affect the calculation of the metric
demo <- GRangesList()
demo[["sample01"]] <- GRanges(seqnames = "chr1", 
    ranges =  IRanges(start = c(1905048, 4554832, 31686841), 
    end = c(2004603, 4577608, 31695808)), strand =  "*",
    state = c("AMPLIFICATION", "AMPLIFICATION", "DELETION"))

demo[["sample02"]] <- GRanges(seqnames = "chr1", 
    ranges =  IRanges(start = c(1995066, 31611222, 31690000), 
    end = c(2204505, 31689898, 31895666)), strand =  c("-", "+", "+"),
    state = c("AMPLIFICATION", "AMPLIFICATION", "DELETION"))

## The amplified region in sample03 is a subset of the amplified regions 
## in sample01
demo[["sample03"]] <- GRanges(seqnames = "chr1", 
    ranges =  IRanges(start = c(1906069, 4558838), 
    end = c(1909505, 4570601)), strand =  "*",
    state = c("AMPLIFICATION", "DELETION"))

## Calculating Sorensen metric
calculateOverlapMetric(demo, method="sorensen")

#> CNV Metric done on overlapping regions
#> Metric:
#> sorensen
#> 
#> AMPLIFICATION:
#>            sample01 sample02 sample03
#> sample01         NA       NA       NA
#> sample02 0.04647582       NA       NA
#> sample03 0.05465532        0       NA
#> 
#> 
#> DELETION:
#>           sample01 sample02 sample03
#> sample01        NA       NA       NA
#> sample02 0.0541291       NA       NA
#> sample03 0.0000000        0       NA
#> 

## Calculating Szymkiewicz-Simpson metric
calculateOverlapMetric(demo, method="szymkiewicz")

#> CNV Metric done on overlapping regions
#> Metric:
#> szymkiewicz
#> 
#> AMPLIFICATION:
#>            sample01 sample02 sample03
#> sample01         NA       NA       NA
#> sample02 0.07796751       NA       NA
#> sample03 1.00000000        0       NA
#> 
#> 
#> DELETION:
#>           sample01 sample02 sample03
#> sample01        NA       NA       NA
#> sample02 0.6477475       NA       NA
#> sample03 0.0000000        0       NA
#>