Calculate Weighted Euclidean distance metric between samples.

The weighted Euclidean distance metric corresponds to the euclidean distance between 2 samples multiplied by the natural logarithm of the number of bases of the analyzed segment. The final metric is the squared sum of the values obtained for all segments that are not excluded of the analysis.

calculateWeightedEuclideanDistance(segmentData, minThreshold = 0.2)

Arguments

segmentData
a list marked as a preMetricSegments class that contains the segment information from at least 2 segment files.

minThreshold
a single numeric setting the minimum value to consider two segments as different durant the metric calculation. If the absolute difference is bellow or equal to threshold, the value will be replaced by zero. Default: 0.2.

Value

a matrix TODO

Details

The weighted euclidean distance is \((\sum((x_i - y_i)^2 * log(nbrBases_i))^0.5\) where x and y are the values of 2 samples for a specific segment i and nbrBases the number of bases of the segment i.

Author

Astrid Deschênes, Pascal Belleau

Examples

## TODO

# Path to the exclusion BED file
bedFile <- system.file("extdata", "mm10.blacklist.bed", 
    package = "CNVMetrics")
    
# Path to the directory containing the segmentation files
segDir <- system.file("extdata", package="CNVMetrics")

 ## TODO