Read segment files and their values and exclusion file to generate a dataset that can be used to calculate the metrics.

The segments and their values are extracted from each file, with a "SEG" or "seg" extension, present in the specified directory. Each segment is associated to its original file by having a source metadata field assigned. When specified, the segments for a BED file are loaded and used to exclude regions from further analysis. All segments are gathered together and a disjoin operation is done to create a collection of non-overlapping ranges. Those ranges are going to be used to calculate the metrics.

prepareInformation(
  segDirectory,
  chrInfo,
  bedExclusionFile = NULL,
  segmentWithHeader = FALSE
)

Arguments

segDirectory
a character string, the path to the directory containing the segment files to compare. Only files with the ".seg" (or ".SEG") extension will be used. At least 2 segment files are needed to be able to used the metrics.

chrInfo
a Seqinfo containing the name and the length of the chromosomes to analyze. Only the chromosomes contained in this Seqinfo will be analyzed.

bedExclusionFile
a character string or NULL, the path to the BED file that contains the regions that must be excluded from the analysis. If NULL, it is assumed that there is no region of exclusion for the calculation of the metrics. Default: NULL.

segmentWithHeader
a logical, when TRUE, the segment files have all a header that should not be imported. Default: FALSE.

Value

a list marked as a preMetricSegments class containing a GRanges with the segment information from all segment files.

Details

The supported BED format requires that the first three columns are:

1. chrom - The name of the chromosome.

2. chromStart - The starting position of the feature in the chromosome or scaffold. The first base in a chromosome is numbered 0.

3. chromEnd - The ending position of the feature in the chromosome or scaffold. The chromEnd base is not included in the display of the feature.

Author

Astrid Deschenes, Pascal Belleau

Examples

# Path to the exclusion BED file
bedFile <- system.file("extdata", "mm10.blacklist.bed", 
    package = "CNVMetrics")

## Prepare information about the chromosomes
## The files are from mouse samples, chr1 to chr4
require(GenomeInfoDb)

## Get the information for mouse genome mm10
## Limit the information to chromosomes 1 to 4
mouseInfo <- Seqinfo(genome="mm10")    
mouseInfo <- mouseInfo[c("chr1", "chr2", "chr3", "chr4"),]
            
## Path to the directory containing the segmentation files
segDir <- system.file("extdata", package="CNVMetrics")

# TODO

prepareInformation(segDirectory = segDir, chrInfo = mouseInfo,
    bedExclusionFile = bedFile, segmentWithHeader = TRUE)

#> $segments
#> GRanges object with 80 ranges and 4 metadata columns:
#>        seqnames              ranges strand |  included     demo01     demo02
#>           <Rle>           <IRanges>  <Rle> | <logical>  <numeric>  <numeric>
#>    [1]     chr1     3222631-3527630      * |      TRUE  0.0507345  0.0122252
#>    [2]     chr1    3527631-24612620      * |      TRUE  0.0507345  0.0122252
#>    [3]     chr1   24612621-24612850      * |     FALSE  0.0507345  0.0122252
#>    [4]     chr1   24612851-48881430      * |      TRUE  0.0507345  0.0122252
#>    [5]     chr1   48881431-48881690      * |     FALSE  0.0507345  0.0122252
#>    ...      ...                 ...    ... .       ...        ...        ...
#>   [76]     chr4   70378321-93689534      * |      TRUE -0.0121036 -0.0411124
#>   [77]     chr4  93689535-111602744      * |      TRUE -0.7418605 -0.0411124
#>   [78]     chr4 111602745-118548460      * |      TRUE -0.0894143 -0.0411124
#>   [79]     chr4 118548461-118548700      * |     FALSE -0.0894143 -0.0411124
#>   [80]     chr4 118548701-156508116      * |      TRUE -0.0894143 -0.0411124
#>            demo03
#>         <numeric>
#>    [1]         NA
#>    [2] 0.00622038
#>    [3] 0.00622038
#>    [4] 0.00622038
#>    [5] 0.00622038
#>    ...        ...
#>   [76] -0.0343646
#>   [77] -0.0343646
#>   [78] -0.0343646
#>   [79] -0.0343646
#>   [80] -0.0343646
#>   -------
#>   seqinfo: 4 sequences from an unspecified genome; no seqlengths
#> 
#> $regression
#> $regression[[1]]
#> $regression[[1]]$lm
#> 
#> Call:
#> lm(formula = "y ~ x", data = subData)
#> 
#> Coefficients:
#> (Intercept)            x  
#>     0.05055      4.86746  
#> 
#> 
#> $regression[[1]]$y_used
#> [1] "demo01"
#> 
#> $regression[[1]]$x_used
#> [1] "demo02"
#> 
#> 
#> $regression[[2]]
#> $regression[[2]]$lm
#> 
#> Call:
#> lm(formula = "y ~ x", data = subData)
#> 
#> Coefficients:
#> (Intercept)            x  
#>    -0.04454      0.93860  
#> 
#> 
#> $regression[[2]]$y_used
#> [1] "demo01"
#> 
#> $regression[[2]]$x_used
#> [1] "demo03"
#> 
#> 
#> 
#> $regressedData
#> GRanges object with 80 ranges and 4 metadata columns:
#>        seqnames              ranges strand |  included     demo01    demo02
#>           <Rle>           <IRanges>  <Rle> | <logical>  <numeric> <numeric>
#>    [1]     chr1     3222631-3527630      * |      TRUE  0.0507345  0.110059
#>    [2]     chr1    3527631-24612620      * |      TRUE  0.0507345  0.110059
#>    [3]     chr1   24612621-24612850      * |     FALSE  0.0507345  0.110059
#>    [4]     chr1   24612851-48881430      * |      TRUE  0.0507345  0.110059
#>    [5]     chr1   48881431-48881690      * |     FALSE  0.0507345  0.110059
#>    ...      ...                 ...    ... .       ...        ...       ...
#>   [76]     chr4   70378321-93689534      * |      TRUE -0.0121036  -0.14956
#>   [77]     chr4  93689535-111602744      * |      TRUE -0.7418605  -0.14956
#>   [78]     chr4 111602745-118548460      * |      TRUE -0.0894143  -0.14956
#>   [79]     chr4 118548461-118548700      * |     FALSE -0.0894143  -0.14956
#>   [80]     chr4 118548701-156508116      * |      TRUE -0.0894143  -0.14956
#>            demo03
#>         <numeric>
#>    [1]         NA
#>    [2] -0.0386978
#>    [3] -0.0386978
#>    [4] -0.0386978
#>    [5] -0.0386978
#>    ...        ...
#>   [76] -0.0767908
#>   [77] -0.0767908
#>   [78] -0.0767908
#>   [79] -0.0767908
#>   [80] -0.0767908
#>   -------
#>   seqinfo: 4 sequences from an unspecified genome; no seqlengths
#> 
#> attr(,"class")
#> [1] "preMetricSegments"