Calculate Jaccard metric

Calculate Jaccard metric using overlapping regions between two samples.

Usage,

calculateJaccard(sample01, sample02)

Arguments

sample01

a GRanges which contains a collection of genomic ranges representing copy number events for the first sample.

sample02

a GRanges which contains a collection of genomic ranges representing copy number events for the second sample.

Value

a numeric, the value of the Jaccard metric. If the metric cannot be calculated, NA is returned.

Details

The method calculates the Jaccard metric using overlapping regions between the samples. All regions present in both samples are used for the calculation of the metric.

The Jaccard metric is calculated by dividing the size of the intersection by the size of the union of the two sets. If the the size of the union of the two sets is zero; the value NA is returned instead. The strand of the regions is not taken into account while calculating the intersection.

References

Jaccard, P. (1912), The Distribution of the Flora in the Alpine Zone. New Phytologist, 11: 37-50. DOI: https://doi.org/10.1111/j.1469-8137.1912.tb05611.x

Author

Astrid Deschênes

Examples

## Load required package to generate the two samples
require(GenomicRanges)
#> Loading required package: GenomicRanges
#> Loading required package: stats4
#> Loading required package: BiocGenerics
#> Loading required package: parallel
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, setdiff, sort, table,
#>     tapply, union, unique, unsplit, which.max, which.min
#> Loading required package: S4Vectors
#> 
#> Attaching package: ‘S4Vectors’
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> Loading required package: GenomeInfoDb

## Generate two samples with identical sequence levels
sample01 <- GRanges(seqnames="chr1", 
    ranges=IRanges(start=c(1905048, 4554832, 31686841), 
    end=c(2004603, 4577608, 31695808)), strand="*")
sample02 <- GRanges(seqnames="chr1", 
    ranges=IRanges(start=c(1995066, 31611222), 
    end=c(2204505, 31689898)), strand="*")

## Calculate Sorensen metric    
CNVMetrics:::calculateJaccard(sample01=sample01, sample02=sample02)
#> [1] 0.03096194