Generate a simulated chromosome based on a reference sample

The function generates a list of simulated segments that represent a simulated chromosome based on a reference sample specified by the user. The function only accounts for the positions where a segment is assigned. In addition, the total number of segments is preserved. A Dirichlet distribution is used to assigned new sizes to the segments with respect to the relative initial size of the segment. Then, those new segments are shuffled without replacement. The positions are replaced by values between zero and one that represent the relative position in a chromosome where positions without segment have been removed. To ensure valuable results, the reference sample should have segments covering a good proportion of the chromosome; those should include NEUTRAL segments.

simChr(curSample, chrCur, nbSim)

Arguments

curSample

a GRanges that contains a collection of genomic ranges representing copy number events, including amplified/deleted status, from exactly one sample. The sample must have a metadata column called 'state' with a state, in an character string format, specified for each region (ex: DELETION, LOH, AMPLIFICATION, NEUTRAL, etc.) and a metadata column called 'CN' that contains the log2 copy number ratios.

chrCur

a character string representing the name of the chromosome that is used as reference for the simulation.

nbSim

a single positive integer which is corresponding to the number of simulations that will be generated.

Value

a list containing one entry per simulation. Each entry is a data.frame containing shuffled segments with 6 columns:

ID The name of the simulation.

chr The name fo the chromosome.

start The starting position of the segment; the positions are between zero and one. The segment width is representing the proportional size of the segment relative to the global segment size.

end The ending position of the segment; the positions are between zero and one. The segment width is representing the proportional size of the segment relative to the global segment size.

log2ratio The log2 copy number ratio assigned to the segment.

state The state of the region (ex: DELETION, LOH, AMPLIFICATION, NEUTRAL, etc.).

Author

Astrid Deschênes, Pascal Belleau

Examples

## Load required package to generate the samples
require(GenomicRanges)

## Create one 'demo' genome with 2 chromosomes
## in a GRanges object
## The stand of the regions doesn't affect the calculation of the metric
sample01 <- GRanges(seqnames=c(rep("chr1", 4), rep("chr2", 3)),
    ranges=IRanges(start=c(1905048, 4554832, 31686841, 32686222,
        1, 120331, 725531),
    end=c(2004603, 4577608, 31695808, 32689222, 117121,
        325555, 1225582)),
    strand="*",
    state=c("AMPLIFICATION", "NEUTRAL", "DELETION", "LOH",
        "DELETION", "NEUTRAL", "NEUTRAL"),
    log2ratio=(c(0.5849625, 0, -1, -1, -0.87777, 0, 0)))

## Generates 10 simulated chromosomes (one chromosome per simulated sample)
## based on chromosome 2 from the input sample.
## The shuffled chromosomes have a start and an end between 0 an 1
CNVMetrics:::simChr(curSample=sample01, chrCur="chr2", nbSim=10)
#> [[1]]
#>   ID  chr     start       end log2ratio    state
#> 1 S1 chr2 0.0000000 0.5868716   0.00000  NEUTRAL
#> 2 S1 chr2 0.5868716 0.7497367  -0.87777 DELETION
#> 3 S1 chr2 0.7497367 1.0000000   0.00000  NEUTRAL
#> 
#> [[2]]
#>   ID  chr     start       end log2ratio    state
#> 1 S2 chr2 0.0000000 0.5772400   0.00000  NEUTRAL
#> 2 S2 chr2 0.5772400 0.7555855  -0.87777 DELETION
#> 3 S2 chr2 0.7555855 1.0000000   0.00000  NEUTRAL
#> 
#> [[3]]
#>   ID  chr     start       end log2ratio    state
#> 1 S3 chr2 0.0000000 0.5961384   0.00000  NEUTRAL
#> 2 S3 chr2 0.5961384 0.8349035   0.00000  NEUTRAL
#> 3 S3 chr2 0.8349035 1.0000000  -0.87777 DELETION
#> 
#> [[4]]
#>   ID  chr     start       end log2ratio    state
#> 1 S4 chr2 0.0000000 0.6103614   0.00000  NEUTRAL
#> 2 S4 chr2 0.6103614 0.7539123  -0.87777 DELETION
#> 3 S4 chr2 0.7539123 1.0000000   0.00000  NEUTRAL
#> 
#> [[5]]
#>   ID  chr     start       end log2ratio    state
#> 1 S5 chr2 0.0000000 0.6177848   0.00000  NEUTRAL
#> 2 S5 chr2 0.6177848 0.8719975   0.00000  NEUTRAL
#> 3 S5 chr2 0.8719975 1.0000000  -0.87777 DELETION
#> 
#> [[6]]
#>   ID  chr     start       end log2ratio    state
#> 1 S6 chr2 0.0000000 0.2650809   0.00000  NEUTRAL
#> 2 S6 chr2 0.2650809 0.4104971  -0.87777 DELETION
#> 3 S6 chr2 0.4104971 1.0000000   0.00000  NEUTRAL
#> 
#> [[7]]
#>   ID  chr     start       end log2ratio    state
#> 1 S7 chr2 0.0000000 0.2337165   0.00000  NEUTRAL
#> 2 S7 chr2 0.2337165 0.8557633   0.00000  NEUTRAL
#> 3 S7 chr2 0.8557633 1.0000000  -0.87777 DELETION
#> 
#> [[8]]
#>   ID  chr    start      end log2ratio    state
#> 1 S8 chr2 0.000000 0.146372  -0.87777 DELETION
#> 2 S8 chr2 0.146372 0.757221   0.00000  NEUTRAL
#> 3 S8 chr2 0.757221 1.000000   0.00000  NEUTRAL
#> 
#> [[9]]
#>   ID  chr     start       end log2ratio    state
#> 1 S9 chr2 0.0000000 0.2494194   0.00000  NEUTRAL
#> 2 S9 chr2 0.2494194 0.4034154  -0.87777 DELETION
#> 3 S9 chr2 0.4034154 1.0000000   0.00000  NEUTRAL
#> 
#> [[10]]
#>    ID  chr     start       end log2ratio    state
#> 1 S10 chr2 0.0000000 0.5883417   0.00000  NEUTRAL
#> 2 S10 chr2 0.5883417 0.7374191  -0.87777 DELETION
#> 3 S10 chr2 0.7374191 1.0000000   0.00000  NEUTRAL
#> 

## Generates 4 simulated chromosomes (one chromosome per simulated sample)
## based on chromosome 1 from the input sample.
## The shuffled chromosomes have a start and an end between 0 an 1
CNVMetrics:::simChr(curSample=sample01, chrCur="chr1", nbSim=4)
#> [[1]]
#>   ID  chr     start       end  log2ratio         state
#> 1 S1 chr1 0.0000000 0.1648300  0.0000000       NEUTRAL
#> 2 S1 chr1 0.1648300 0.2345833 -1.0000000      DELETION
#> 3 S1 chr1 0.2345833 0.2619618 -1.0000000           LOH
#> 4 S1 chr1 0.2619618 1.0000000  0.5849625 AMPLIFICATION
#> 
#> [[2]]
#>   ID  chr      start        end  log2ratio         state
#> 1 S2 chr1 0.00000000 0.06837575 -1.0000000      DELETION
#> 2 S2 chr1 0.06837575 0.23422585  0.0000000       NEUTRAL
#> 3 S2 chr1 0.23422585 0.97143751  0.5849625 AMPLIFICATION
#> 4 S2 chr1 0.97143751 1.00000000 -1.0000000           LOH
#> 
#> [[3]]
#>   ID  chr     start       end  log2ratio         state
#> 1 S3 chr1 0.0000000 0.1598338  0.0000000       NEUTRAL
#> 2 S3 chr1 0.1598338 0.1838245 -1.0000000           LOH
#> 3 S3 chr1 0.1838245 0.2525502 -1.0000000      DELETION
#> 4 S3 chr1 0.2525502 1.0000000  0.5849625 AMPLIFICATION
#> 
#> [[4]]
#>   ID  chr     start       end  log2ratio         state
#> 1 S4 chr1 0.0000000 0.7580155  0.5849625 AMPLIFICATION
#> 2 S4 chr1 0.7580155 0.8036813 -1.0000000      DELETION
#> 3 S4 chr1 0.8036813 0.9770889  0.0000000       NEUTRAL
#> 4 S4 chr1 0.9770889 1.0000000 -1.0000000           LOH
#>